refactor : point/infrastructure 폴더 밑으로 이동
