+++ src/main/java/com/kakaotech/team14backend/inner/point/usecase/GetPointUsecase.java
@@ -5,7 +5,7 @@ import com.kakaotech.team14backend.exception.MemberNotFoundException;
 import com.kakaotech.team14backend.inner.member.model.Member;
 import com.kakaotech.team14backend.point.domain.GetPointPolicy;
 import com.kakaotech.team14backend.point.domain.Point;
-import com.kakaotech.team14backend.inner.point.repository.PointRepository;
+import com.kakaotech.team14backend.point.infrastructure.PointRepository;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
