+++ src/main/java/com/kakaotech/team14backend/point/infrastructure/PointRepository.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.inner.point.repository;
+package com.kakaotech.team14backend.point.infrastructure;
 
 import com.kakaotech.team14backend.point.domain.Point;
 import org.springframework.data.jpa.repository.JpaRepository;
